requirements.txt
python-dotenv==1.0.1
selenium==4.19.0
google-cloud-storage==2.16.0